+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 
 ####2014-7-3
 + 部分实现了 resolvePromise 方法，但是其中如果 x 为 object 或者 function 时如何处理则还没有看懂promise/a+ 的 spec。
+
+####2014-07-15
++ 功能完成，在运行 promise aplus test 的时候 console 产生太多的日志没法看错误，需要寻找一种看的方法